Add backticks around shell command
This way you can uncomment it and run it from the editor.

Useful b/c the shell doesn't necessarily have the same Ruby env as the editor,
so it's simpler to gem install with backticks than to have to think through
which Ruby SiB is configured for and how to get into that environment to run
the command.